### 2024-03-12 #6471 @editorialbot
[PRE REVIEW]: CoastalLens: A MATLAB UAV Video Stabilization and Rectification Framework
**Submitting author:** @AthinaLange (<a href="http://orcid.org/0000-0002-7650-9976">Athina M.Z. Lange</a>)
**Repository:** https://github.com/AthinaLange/CoastalLens
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
