### 2024-03-12 #6471 @kthyng
Hi @AthinaLange! I am checking your submission for a few things to make sure it can move forward into review.

* I see there is a demo which can be used to verify if the software runs properly ✅  — however the data is available on a google drive link. This should be available somewhere permanent. How big is that file? Can you make it any smaller while keeping its utility as a test file? Depending on its size you could potentially keep it in the github repo itself, otherwise figshare could be a good place to stash it, or maybe UCSD has a place for it. It needs to be somewhere more permanent than a google drive link.
* I see installation instructions ✅
* docs — I see a directory called docs — Where can I find the docs?
