### 2024-03-12 #6471 @AthinaLange
Hi @kthyng,

- because of the camera calibration, the video can't be compressed. I have uploaded the file to Zenodo and inserted the new link there.

- The 'docs' directory is where all the figures for the wiki are stored. The majority of the documentation is in the [wiki](https://github.com/AthinaLange/CoastalLens/wiki).
This software will be added to the CIRN (Coastal Imaging Research Network) repositories and having the documentation in a wiki has become standard practice there.

Thanks
