### 2024-03-12 178273c
Update extract_images_from_UAV.m
