### 2024-03-20 #6471 @kthyng
@hugoledoux Ok fair enough. I will see if someone in image processing becomes available in the meantime. If not, perhaps you'll be hearing from me again! 😆  Enjoy your holidays!
