+++ CODES/helperFunctions/getCoords_DEM.m
@@ -1,0 +1,120 @@
+function [xyz, localX, localY, Z, Eastings, Northings] = getCoords_DEM(Products, DEM)
+% getCoords returns the (x,y,z) world coordinates for the dimensions specified in Products.
+%% Syntax
+%            [xyz] = getCoords(Products)
+%
+%% Description
+%   Args:
+%           Products (structure) : Single Products object. All necessary variables given in define_grid, define_xtransect, or define_ytransect
+%                       type (string) : 'Grid', 'xTransect', 'yTransect'
+%                       frameRate (double) : frame rate of product (Hz)
+%                       lat (double) : latitude of origin grid
+%                       lon (double): longitude of origin grid
+%                       angle (double): shorenormal angle of origid grid (degrees CW from North)
+%                       xlim (double): [1 x 2] cross-shore limits of grid (+ is offshore of origin) (m)
+%                       ylim (double) : [1 x 2] along-shore limits of grid (+ is to the right of origin looking offshore) (m)
+%                       dx (double) : Cross-shore resolution (m)
+%                       dy (double) : Along-shore resolution (m)
+%                       x (double): Cross-shore distance from origin (+ is offshore of origin) (m)
+%                       y (double): Along-shore distance from origin (+ is to the right of the origin looking offshore) (m)
+%                       z (double) : Elevation - can be empty or array of DEM values (NAVD88 m)
+%                       tide (double) : Tide level (NAVD88 m)
+%
+%   Returns:
+%       xyz (double) : [m x 3] (x, y, z) world coordinates for Products. Rotated according to Product.angle.
+%       localX (double) : [y x x] local X coordinates (+x is offshore, m)
+%       localY (double) : [y x x] local Y coordinates (+y is right of origin, m)
+%       Z (double) : [y x x] Z coordinate - tide level
+%       Eastings (double) : [y x x] Eastings coordinates (m)
+%       Northings (double) : [y x x] Northings coordinates (m)
+%
+%
+%% Citation Info
+% github.com/AthinaLange/CoastalLens
+% Jan 2024;
+
+%% Data
+assert(isa(Products, 'struct'), 'Error (getCoords): Products must be a structure.')
+assert(size(Products,2)==1, 'Error (getCoords): Products must be a single object structure. Pass as Products(pp).')
+assert(isfield(Products, 'lat'), 'Error (getCoords): Products must have lat field.')
+assert(isfield(Products, 'lon'), 'Error (getCoords): Products must have lon field.')
+assert(isfield(Products, 'angle'), 'Error (getCoords): Products must have angle field.')
+assert(isfield(Products, 'tide'), 'Error (getCoords): Products must have tide field.')
+assert(isfield(Products, 'xlim'), 'Error (getCoords): Products must have xlim field.')
+assert(isfield(Products, 'ylim'), 'Error (getCoords): Products must have ylim field.')
+assert(isfield(Products, 'type'), 'Error (getCoords): Products must have type field.')
+assert(isfield(Products, 'x'), 'Error (getCoords): Products must have x field.')
+assert(isfield(Products, 'y'), 'Error (getCoords): Products must have y field.')
+
+%%
+% Get origin coordinates
[…]
