### 2024-04-30 #6471 @Kevin-Mattheus-Moerman
@simmonsja would you be interested in reviewing this work for JOSS? It features video stabilisation of coastal video footage.

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06471/joss.06471/10.21105.joss.06471.pdf
Software: https://github.com/AthinaLange/CoastalLens
