### 2024-05-07 #6471 @simmonsja
Hi @Kevin-Mattheus-Moerman,
Unfortunately I don't have access to MATLAB software with my new organisation, so hard to review this submission.
@kvos may be a more suitable reviewer if he has capacity and has quite a bit of experience with UAV.
Cheers,
Josh
