### 2024-05-30 #6471 @Kevin-Mattheus-Moerman
@mitchharley @tdprice @KristenSplinter @chrisleaman @eloraoades @JennaBrown-USGS @eswanson-usgs @kerfoot would you be interested in reviewing this work for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)? It features software for video stabilisation of coastal video footage.

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06471/joss.06471/10.21105.joss.06471.pdf
Software: https://github.com/AthinaLange/CoastalLens

JOSS reviews focus on evaluation of the software as well as the review of the very short paper. JOSS reviews happen in the open here on GitHub. See here for more information on our review process: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

If you are interested you can let me know here. Thanks!
