### 2024-05-30 #6471 @KristenSplinter
Hi Kevin,
I’m not available currently.
Meg palmsten at USGS would be a possibility.
Meg Palmsten ***@***.***>

Kristen D. Splinter (she/her)
Associate Professor, Managing Director Water Research Laboratory
ARC Future Fellow

Water Research Laboratory
School of Civil and Environmental Engineering, UNSW Sydney
110 King St, Manly Vale NSW 2093 Australia

T: +(02) 8071 9845**please note the new number
W: www.wrl.unsw.edu.au<http://www.wrl.unsw.edu.au/>

[A picture containing text  Description automatically generated]<http://www.wrl.unsw.edu.au/>
CRICOS Provider Code 00098G

Check out my latest publications:

  *   Vos, K., M. Harley, I. Turner, K.Splinter (2023). Pacific shoreline erosion and accretion patterns controlled by El Niño/Southern Oscillation. Nature Geoscience 16 (2), 140-146. https://www.nature.com/articles/s41561-022-01117-8
  *   Vitousek, S.,  K Vos, KD Splinter, L Erikson, PL Barnard (2023). A model integrating satellite-derived shoreline observations for predicting fine-scale shoreline response to waves and sea-level rise across large coastal regions. Journal of Geophysical Research: Earth Surface, e2022JF006936. doi: 10.1029/2022JF006936
  *   Ibaceta, R.,  MD Harley, IL Turner, KD Splinter (2023) Interannual variability in dominant shoreline behaviour at an embayed beach. Geomorphology 433, 108706 https://doi.org/10.1016/j.geomorph.2023.108706
  *   Martins, K., P. Bonneton, O. de Viron, I.L. Turner, M.D. Harley, K. Splinter (2023). New Perspectives for Nonlinear Depth-Inversion of the Nearshore Using Boussinesq Theory. http://doi.org/10.1029/2022GL100498
  *   Ibaceta, R., K.D. Splinter, M.D. Harley, I.L. Turner (2022). Improving multi-decadal coastal shoreline change predictions by including model parameter non-stationarity. https://doi.org/10.3389/fmars.2022.1012041

A full list of my publications can be found here: https://scholar.google.com/citations?user=DsXnQcUAAAAJ&hl=en

From: Kevin Mattheus Moerman ***@***.***>
Reply to: openjournals/joss-reviews ***@***.***>
Date: Thursday 30 May 2024 at 6:39 AM
Cc: Kristen Splinter ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: CoastalLens: A MATLAB UAV Video Stabilization and Rectification Framework (Issue #6471)

@mitchharley<https://github.com/mitchharley> @tdprice<https://github.com/tdprice> @KristenSplinter<https://github.com/KristenSplinter> @chrisleaman<https://github.com/chrisleaman> @eloraoades<https://github.com/eloraoades> @JennaBrown-USGS<https://github.com/JennaBrown-USGS> @eswanson-usgs<https://github.com/eswanson-usgs> @kerfoot<https://github.com/kerfoot> would you be interested in reviewing this work for The Journal of Open Source Software (JOSS)<https://joss.theoj.org/>? It features software for video stabilisation of coastal video footage.

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06471/joss.06471/10.21105.joss.06471.pdf
Software: https://github.com/AthinaLange/CoastalLens

JOSS reviews focus on evaluation of the software as well as the review of the very short paper. JOSS reviews happen in the open here on GitHub. See here for more information on our review process: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

If you are interested you can let me know here. Thanks!
