+++ CODES/scripts/get_products.m
@@ -1,0 +1,246 @@
+%% get_products
+% get_products returns extracted image pixel for coordinates of Products and saves Timex, Brightest and Darkest image products
+%% Description
+%
+%   Inputs:
+%           global_dir (string) : global directory - where CODES and (typically) DATA  are located.
+%           day_files (structure) : folders of the days to process - requires day_files.folder and day_files.name
+%           flights (structure) : folders of the flights to process - requires flights.folder and flights.name
+%           extract_Hz (double) : extraction frame rate (Hz) - obtained from Products
+%           R (structure) : extrinsics/intrinsics information
+%                       intrinsics (cameraIntrinsics) : camera intrinsics as calibrated in the cameraCalibrator tool
+%                       extrinsics_2d (projtform2d) : [1 x m] 2d projective transformation of m images (with respect to 1st frame)
+%                       worldPose (rigidtform3d) : orientation and location of camera in world coordinates, based off ground control location (pose, not extrinsic)
+%                       t (datetime array) : [1 x m] datetime of images at various extraction rates in UTC
+%           Products (structure) : Data products
+%                       type (string) : 'Grid', 'xTransect', 'yTransect'
+%                       frameRate (double) : frame rate of product (Hz)
+%                       lat (double) : latitude of origin grid
+%                       lon (double): longitude of origin grid
+%                       angle (double): shorenormal angle of origid grid (degrees CW from North)
+%                       xlim (double): [1 x 2] cross-shore limits of grid (+ is offshore of origin) (m)
+%                       ylim (double) : [1 x 2] along-shore limits of grid (+ is to the right of origin looking offshore) (m)
+%                       dx (double) : Cross-shore resolution (m)
+%                       dy (double) : Along-shore resolution (m)
+%                       x (double): Cross-shore distance from origin (+ is offshore of origin) (m)
+%                       y (double): Along-shore distance from origin (+ is to the right of the origin looking offshore) (m)
+%                       z (double) : Elevation - can be empty or array of DEM values (NAVD88 m)
+%                       tide (double) : Tide level (NAVD88 m)
+%
+%   Returns:
+%           Products (structure) : Data products
+%                       type (string) : 'Grid', 'xTransect', 'yTransect'
+%                       frameRate (double) : frame rate of product (Hz)
+%                       lat (double) : latitude of origin grid
+%                       lon (double): longitude of origin grid
+%                       angle (double): shorenormal angle of origid grid (degrees CW from North)
+%                       xlim (double): [1 x 2] cross-shore limits of grid (+ is offshore of origin) (m)
+%                       ylim (double) : [1 x 2] along-shore limits of grid (+ is to the right of origin looking offshore) (m)
+%                       dx (double) : Cross-shore resolution (m)
+%                       dy (double) : Along-shore resolution (m)
+%                       x (double): Cross-shore distance from origin (+ is offshore of origin) (m)
+%                       y (double): Along-shore distance from origin (+ is to the right of the origin looking offshore) (m)
+%                       z (double) : Elevation - can be empty or array of DEM values (NAVD88 m)
+%                       tide (double) : Tide level (NAVD88 m)
[…]
