+++ random_play_codes/UAV_CIRN_cBathy_full_codes_vold.m
@@ -1,555 +1,0 @@
-%% UAV CIRN & timestack Workflow
-
-aa=split(data_dir,'/');
-stationStr = char(aa(end))
-close all 
-clearvars -except stationStr *_dir
-
-%% ========================================================================
-%                                   CIRN
-% =========================================================================
-
-%% Loop through all flights from day - extract images
-for nn = 1:num_flights
-    %% Define file location
-    
-    video_path = fullfile(data_dir,flights(nn,:))
-    cd(video_path)
-    aa=split(pwd,'/');
-    filename = char(aa(end))
-    mkdir('Processed_data')
-    %% FILE NAMES
-    odir = pwd;
-    oname = stationStr;
-    csvname = [filename '.csv'];
-    stackName = [ stationStr '_' filename];
-    
-     %% A0 - Image Extraction - 10Hz
-    % Location where images will be saved
-    tic
-    
-    imageDirectory = 'images_10Hz/';
-    %if ~exist(imageDirectory) % extract 10Hz images
-        mkdir(imageDirectory)
-        for ii = 1:length(mov_id)
-            mkdir(['images_10Hz/' char(string(ii))])
-            system(['ffmpeg -i ' char(string(C.FileName(mov_id(ii)))) ' -qscale:v 2 -r 10 images_10Hz/' char(string(ii)) '/Frame_%04d.jpg'])
-        end
-        for ii = 1:length(mov_id)
-            Ltemp=sort(split(string(ls(imageDirectory)))); Ltemp=Ltemp(2+length(mov_id):end); % Athina: changes
-            Lfull = length(Ltemp);
-            L=sort(split(string(ls([imageDirectory char(string(ii))])))); L=L(2:end); % Athina: changes
-            if ii == 1
-                movefile(sprintf('images_10Hz/%d/Frame_*', ii),'images_10Hz/')
-            else
-                for ll = 1:length(L)
-                    if ll < 10
-                        id = ['000' char(string(ll))];
-                    elseif ll < 100
-                        id = ['00' char(string(ll))];
-                    elseif ll < 1000
-                        id = ['0' char(string(ll))];
-                    else
-                        id = [char(string(ll))];
-                    end
-            
-                    movefile(sprintf('images_10Hz/%d/Frame_%s.jpg', ii,id),sprintf('images_10Hz/Frame_%s.jpg', char(string(ll+Lfull))))
-                end
-            end
-        end
-    
-        for ii = 1:length(mov_id); rmdir(sprintf('images_10Hz/%d', ii), 's');end
-   % end
-    cd('..')
-    clearvars -except num_flights flights *_dir stationStr nn
-
-end
-
-%% Loop through all flights from day - clicking step
-for nn = 1:num_flights
-    %%
-    clearvars -except *dir video_path flights num_flights stationStr nn
-    nn
-    video_path = fullfile(data_dir,flights(nn,:))
-    cd(video_path)
-    aa=split(pwd,'/');
-    filename = char(aa(end))
-    
-    odir = pwd;
-    oname = [ stationStr '_' filename];
-    csvname = [filename '.csv'];
-    stackName = [ stationStr '_' filename];
-
-    %% E - Choosing SCP
-    % click on scp point, give it a number (matches the number in gps.txt), 
[…]
