+++ change_log.txt
@@ -2,4 +2,5 @@ v1.0.01
 
 - added in DEM projection option
 - changed reference image for stabilization to 1st frame
-
+- integrated ARGUS option
+- created cBathy and cBathyCT workflow
