+++ CODES/cBathy/cBathyCT_workflow.m
@@ -1,207 +1,0 @@
-%% cBathyCT workflow
-%   1. Housekeeping
-%           Global Directory Selection: Asks the user to choose the root directory for UAV rectification.
-%           Directory Checking: Checks if required code directories and MATLAB toolboxes exist, prompting the user to download them if not. Adds required code directories to MATLAB's search path
-%           Folder Selection: Allows the user to choose which days to process.
-%   2. run_cBathy
-%           Steps 1 & 2 (no Kalman filtering right now)
-%           Holman and Bergsma 2022
-%   3. save_timestacks
-%           save timestacks as pngs -> then run Wave Crest Detector in terminal
-%   4. run_cbathyCT
-%           Lange et al. (2023)
-%
-%% Description
-%
-%   Inputs: 
-%           DEM
-%           Products - from UAV_rectification or ARGUS_rectification
-%
-%
-%   Returns:
-%           cBathyCT (structure) : cBathyCT results
-%                       date (string) : date of product (YYYYMMDD)
-%                       Iocation (string) : location name - from filename
-%                       flight / cam (double) : flight number (for UAV) or camera number (for ARGUS)
-%                       mop (double) : mop number
-%                       x10 (array) : [0:0.1:500], 0.1m resolution - all elevation values interpolated to this x array
-%                       survey (structure) : 
-%                                   z (array) : [1 x length(x10)] elevation pulled from survey DEM on MOP line
-%                       cbathy (structure): 
-%                                   z (array) : cBathy direct output on given transect
-%                                   zerr (array) : cBathy hErr
-%                                   cbathy_hErr (array) : cBathy with hErr > 0.5m region interpolated over
-%                                   cbathy_gamma (array) : cBathy with breaking region removed for given variable gamma(x)
-%                       tide (double) : tide level - pulled from Products tide level
-%                       min_tide (double) : minimum tide level of day - pulled from DEM - assuming survey taken at daily low tide
-%                       crests (structure) : 
-%                                   t (array) : time for wave tracks (sec)
-%                                   x (array) : cross-shore location for wave tracks (m) (0 is offshore, 500 is onshore)
-%                                   c (array) : phase speed of wave tracks (m/s) - interpolated to x10
-%                       bp (array) : breakpoint location for each crest 
-%                       xshift (double) : index (for x10) of cross-shore shift required to match subaerial survey with subaqueous bathymetry - should be small if using DEM
-%                       h_avg (structure) : 
-%                                   lin (array) : linear crest-tracking c = sqrt(gh)
-%                                   nlin (array) : nonlinear crest-tracking  = sqrt(gh(1+0.42))
-%                                   bp (array) : breakpoint transition crest-tracking c = sqrt(gh(1+gamma(x)))
-%                       gamma (array) : [ x ] transition between 0 and 0.42 for each wave track 0 interpolated to x10 
-%                       gamma_mean (array) : [1 x length(x10)] mean of all step function gammas 
-%                       composite (structure) : constructed topobathy products with subaerial survey + 
-%                                   cbathy_hErr (array) : cBathy with hErr < 0.5m
-%                                   cbathy_gamma (array) : cBathy with breaking region removed
-%                                   cbathy_nlin (array) : cbathy_gamma with gamma(x) correction
-%                                   cbathyCT (array) : breakpoint transition crest-tracking surfzone bathymetry and cBathy offshore
[…]
