### 2024-07-30 2ac371d
new improved method of extracting pixels - speeds it up
