+++ ARGUS_rectification.m
@@ -3,7 +3,7 @@
 %           Global Directory Selection: Asks the user to choose the root directory for UAV rectification.
 %           Directory Checking: Checks if required code directories and MATLAB toolboxes exist, prompting the user to download them if not. Adds required code directories to MATLAB's search path
 %           Folder Selection: Allows the user to choose which days to process.
-%   2. User Input Section
+%   2. User Input Section%%
 %           Camera Intrinsics: User inputs camera intrinsics file and related details.
 %           Products Definition: User inputs or loads grid and transect coordinates.
 %           Initial Camera Extrinsics: Use ground control points to define camera world pose.
@@ -259,10 +259,10 @@ if ~exist('DEM', 'var')
                     disp('Please load in DEM topo file.')
                     [temp_file, temp_file_path] = uigetfile(global_dir, 'DEM topo file');
                     load(fullfile(temp_file_path, temp_file)); clear temp_file*
-                    assert(isfield(DEM, 'time'), 'Error (input_day_flight_data.m): DEM does not have time field.')
-                    assert(isfield(DEM, 'X_gridded'), 'Error (input_day_flight_data.m): DEM does not have X_gridded field.')
-                    assert(isfield(DEM, 'Y_gridded'), 'Error (input_day_flight_data.m): DEM does not have Y_gridded field.')
-                    assert(isfield(DEM, 'Z_gridded'), 'Error (input_day_flight_data.m): DEM does not have Z_gridded field.')
+                    assert(isfield(DEM, 'time'), 'Error (ARGUS_rectification.m): DEM does not have time field.')
+                    assert(isfield(DEM, 'X'), 'Error (ARGUS_rectification.m): DEM does not have X field.')
+                    assert(isfield(DEM, 'Y'), 'Error (ARGUS_rectification.m): DEM does not have Y field.')
+                    assert(isfield(DEM, 'Z'), 'Error (ARGUS_rectification.m): DEM does not have Z field.')
                 case 'No'
                     [DEM] = define_DEM;
                     answer4 = questdlg('Do you want to save this DEM file for the future?', 'Save DEM file', 'Yes', 'No', 'Yes');
@@ -284,70 +284,74 @@ end
 %                             extrinsics, intrinsics, initial frame, input data, products
 %  =====================================================================
 [Products.tide] = deal(0);
-for cc = 1:cam_num
 
-    [xyz,~,~,~,~,~] = getCoords(Products(1));
-    [y2,x2, ~] = ll_to_utm(Products(1).lat, Products(1).lon);
+ids_grid = find(ismember(string({Products.type}), 'Grid'));
+for pp = ids_grid
+    for cc = 1:cam_num
 
-    %aa=xyz-[x2 y2 0];
-    %id_origin=find(min(abs(aa(:,[1 2])))==abs(aa(:,[1 2])));
-    %iP = round(world2img(xyz, pose2extr(R(cc).worldPose), R(cc).cameraParams.Intrinsics));
-    %iP_origin = iP(id_origin);
-    clear aa iP
+        [xyz,~,~,~,~,~] = getCoords(Products(pp));
+        [y2,x2, ~] = ll_to_utm(Products(pp).lat, Products(pp).lon);
 
-    aa=xyz-[R(cc).worldPose.Translation(1) R(cc).worldPose.Translation(2) 0];
-    % This is built in for Fletcher - might not be valid for other sites
-    if Products(1).angle > 180
-        if cc == 1
-            id=[];
-            % behind camera
-            for ii = 1:length(aa)
-                if aa(ii,1) > 0 & aa(ii,2) < 0
-                    id = [id ii];
[…]
