+++ CODES/scripts/extract_images_from_UAV.m
@@ -1,0 +1,118 @@
+%% extract_images_from_UAV
+% extract_images_from_UAV extracts images from video files at specified frame rates for all flights on specified processing days.
+%% Description
+%
+%   Inputs:
+%           global_dir (string) : global directory - where CODES and (typically) DATA  are located.
+%           day_files (structure) : folders of the days to process - requires day_files.folder and day_files.name
+%           flights (structure) : folders of the flights to process - requires flights.folder and flights.name
+%           extract_Hz (double) : extraction frame rate (Hz) - obtained from Products
+%           C (table) : Table of image/video metadata (from get_metadata)
+%           mov_id (array) : list of ids that match videos (MOV, MP4, TS) in metadata table (from find_file_format_id)
+%
+%
+% Requires ffmpeg (https://ffmpeg.org)
+%
+% For each extraction frame rate:
+%           - make Hz directory for images
+%           - for every movie to be extracted: extract images from video at extraction frame rate using ffmpeg (into separate folders initially)
+%           - move images from movie folders into group folder and rename sequentially
+%% Function Dependenies
+% extract_images
+% combine_images
+%
+%% Citation Info
+% github.com/AthinaLange/CoastalLens
+% Sept 2023;
+
+%% Data
+
+if ~exist('global_dir', 'var') || ~exist('day_files', 'var') || ~isstruct(day_files) || ~isfield(day_files, 'folder') || ~isfield(day_files, 'name')
+    disp('Missing global_dir and day_files. Please load in processing_run_DD_Month_YYYY.mat that has the day folders that you would like to process. ')
+    [temp_file, temp_file_path] = uigetfile(pwd, 'processing_run_.mat file');
+    load(fullfile(temp_file_path, temp_file)); clear temp_file*
+    assert(isfolder(global_dir),['Error (extract_images_from_UAV): ' global_dir 'doesn''t exist.']);
+
+    if ~exist('global_dir', 'var')
+        disp('Please select the global directory.')
+        global_dir = uigetdir('.', 'UAV Rectification');
+        cd(global_dir)
+    end
+    if ~exist('day_files', 'var') || ~isstruct(day_files) || ~isfield(day_files, 'folder') || ~isfield(day_files, 'name')
+        disp('Choose DATA folder.')
+        disp('For Athina: DATA')
+        data_dir = uigetdir('.', 'DATA Folder');
+
+        day_files = dir(data_dir); day_files([day_files.isdir]==0)=[]; day_files(contains({day_files.name}, '.'))=[];
+        [ind_datafiles,~] = listdlg('ListString',{day_files.name}, 'SelectionMode','multiple', 'InitialValue',1, 'PromptString', {'Which days would you like to process?'}, 'ListSize', [500 300]);
+        day_files = day_files(ind_datafiles);
+    end
+end % if exist('global_dir', 'var')
+
+% check that needed files exist
+for dd = 1:length(day_files)
+    assert(isfile(fullfile(day_files(dd).folder, day_files(dd).name, 'day_config_file.mat')),['Error (extract_images_from_UAV): ' fullfile(day_files(dd).folder, day_files(dd).name, 'config_file.mat') ' doesn''t exist.']);
+    load(fullfile(day_files(dd).folder, day_files(dd).name, 'day_config_file.mat'), 'flights')
+    for ff = 1:length(flights)
+        assert(isfile(fullfile(flights(ff).folder, flights(ff).name, 'Processed_data', 'Initial_coordinates.mat')), ['Error (extract_images_from_UAV): ' fullfile(flights(ff).folder, flights(ff).name, 'Processed_data', 'Initial_coordinates.mat') ' doesn''t exist.']);
[…]
