### 2024-08-06 9f84084
fix get_products
