### 2024-08-06 #6471 @AthinaLange
@Kevin-Mattheus-Moerman
Matt Kabernagel is a MS student at the Naval Postgraduate School working with Mara Orescanin. Mara suggested he try running CoastalLens on some UAV data they have collected and implement some other work I've done on that data.
I am answering some questions about the other work since it has not been fully documented yet, but I have not discussed CoastalLens with him at all.
