### 2024-08-07 314d429
fix DEM option in plot
