### 2024-08-07 #6471 @Kevin-Mattheus-Moerman
Thanks @kabernagelm. This is still the pre-review issue so we have not formally started reviewing yet (we'll open a dedicated review issue for that), but in the mean time you can familiarize yourself with our guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Secondly, if you could update your GitHub profile with some additional information (full name and affiliation etc) that would be great. Thanks.
