### 2024-08-07 #6471 @kthyng
@burritobrittany or @KateBrodie This submission to the Journal for Open Source Software is related to your package https://github.com/Coastal-Imaging-Research-Network/CIRN-Quantitative-Coastal-Imaging-Toolbox. Would either of you be interested in reviewing it? We hope for the review to be done in 4-6 weeks and it tends to be iterative between the author and reviews, based on a checklist open on github. Thanks for your consideration!
