### 2024-08-08 #6471 @KateBrodie
@kthyng both Brittany and I are likely too close to this work to review it. I would suggest Dr. Matt Conlin (@conlin-matt) a post-doc at OSU, or Dr. Annika O’Dea (@anodea) a Researcher at the US Army Corps of Engineers, or Maile McCann (mailemcc@usc.edu) a Ph.D candidate who has written a similar coastal imaging toolbox in python (but who can also do matlab).
