### 2024-08-13 #6471 @chickadel
@kthyng Sure I could review this.  Never done a JOSS review before, but looks pretty straightforward.
