### 2024-09-11 #7111 @chickadel
In. the demo data, the workflow wasn't clear and I couldn't get past the GCP selection (it keeps reappearing) in the code.  Without a manual/video of how to use the code, I can't proceed.
