### 2024-09-11 #7111 @chickadel
RE: 'State of the field' in the 'paper', I didn;t find a section here, or an explicit statement dedicated to this.  I think the reality is that this is a special case of image stabilization using in-view features, which is common, now, so there should be some references available.
