### 2024-09-13 cd8453e
Update input_day_flight_data_DEMO.m
Added additional comment at GCP section
