### 2024-09-13 #7111 @AthinaLange
@chickadel The worldPose calculation of the camera extrinsics is based on a best fit approximation of the GCP locations. For the demo example, I chose the points that tend (~90% of the time) to provide a worldPose estimate immediately. The code is built to request new points if it cannot determine a pose from the initial coordinates. There is also the option to select 'No more points possible'. This option should by-pass the further requests. I have added in displayed text in the DEMO version to advise users to select the 'No more points possible.' (new line 477).
Image examples of the steps can be found in the wiki [here](https://github.com/AthinaLange/CoastalLens/wiki/select_survey_gcp) and [here](https://github.com/AthinaLange/CoastalLens/wiki/select_target_gcp).

Hopefully this clears up this issue, if not, please let me know if there's a specific error that is getting thrown.
