### 2024-09-13 #7111 @AthinaLange
@chickadel regarding ffmpeg and exiftool, these can be installed from the source code or static builds as you would any command-line application.
