### 2024-09-13 #7111 @AthinaLange
@chickadel regarding the 'state of the field', this was addressed in the 'Statement of need' section. While the use of in-view features for stabilization is indeed widely done for a variety of fields, to our knowledge the only existing workflow/codes that exists for the stabilization, rectification and extraction of relevant products for coastal imagery is the CIRN Qualitative Coastal Imaging Toolbox and tangentially SurfRCaT.
We have included the reference to these toolboxes, as well as numerous references that support the various steps of stabilization and rectification that are done in this workflow.

If you know of any other coastal imagery related stabilization and rectification workflows, we are happy to include it in our references.
