### 2024-09-16 #7111 @chickadel
> @chickadel regarding the 'state of the field', this was addressed in the 'Statement of need' section. While the use of in-view features for stabilization is indeed widely done for a variety of fields, to our knowledge the only existing workflow/codes that exists for the stabilization, rectification and extraction of relevant products for coastal imagery is the CIRN Qualitative Coastal Imaging Toolbox and tangentially SurfRCaT. We have included the reference to these toolboxes, as well as numerous references that support the various steps of stabilization and rectification that are done in this workflow.
>
> If you know of any other coastal imagery related stabilization and rectification workflows, we are happy to include it in our references.

If you limit this to only software for strict nearshore application, yes I agree.  However there's a work of SIFT (and other) based stabilization software out there for use with UAS, and I guess I was referring to that.  Clearly, I'm not expecting a full review here, which is way beyond the scope, but maybe a recognition (with maybe a review article reference?) that this software addresses a corner of the possible applications for this type of work.
