+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 ---
 
 # Statement of need
-Uncrewed aerial vehicles (UAVs) are an important tool for coastal monitoring with their relatively low-cost and rapid deployment capabilities. To generate scientific-grade image products, to use for wave runup observations, for bathymetry inversions, or tracking surfzone currents, the hovering UAV images/videos must be stabilized and rectified into world coordinates. Due to the limited stationary region of coastal images suitable for control points, the processing of  UAV-obtained videos can be time-consuming and resource-intensive. The [CIRN Qualitative Coastal Imagining Toolbox](https://github.com/Coastal-Imaging-Research-Network/CIRN-Quantitative-Coastal-Imaging-Toolbox)[@bruder:2020] provided a first-of-its-kind open-sourced code for rectifying these coastal UAV videos. Limitations of the toolbox, however, prompted the development of CoastalLens with an efficient data input procedure, providing capabilities to obtain drone position (extrinsics) from LiDAR surveys, and using a feature detection and matching algorithm to stabilize the video prior to rectification. This framework reduces the amount of human oversight, now only required during the data input processes. Removing the dependency on threshold stability control points provides more stable results and can also result in less time in the field. We hope this framework will allow for more efficient processing of the ever-increasing coastal UAV datasets. 
+Uncrewed aerial vehicles (UAVs) are an important tool for coastal monitoring with their relatively low-cost and rapid deployment capabilities. To generate scientific-grade image products, to use for wave runup observations, for bathymetry inversions, or tracking surfzone currents, the hovering UAV images/videos must be stabilized and rectified into world coordinates. Video stabilization is an active area of research and is use in many fields including videography, law enforcement, and agriculture (see [@guilluy:2021], [wang:2023] for reviews on video stabilization). Due to the limited stationary region of coastal images suitable for control points, the processing of coastal UAV-obtained videos can be time-consuming and resource-intensive. The [CIRN Qualitative Coastal Imagining Toolbox](https://github.com/Coastal-Imaging-Research-Network/CIRN-Quantitative-Coastal-Imaging-Toolbox)[@bruder:2020] provided a first-of-its-kind open-sourced code for rectifying these coastal UAV videos. Limitations of the toolbox, however, prompted the development of CoastalLens with an efficient data input procedure, providing capabilities to obtain drone position (extrinsics) from LiDAR surveys, and using a feature detection and matching algorithm to stabilize the video prior to rectification. This framework reduces the amount of human oversight, now only required during the data input processes. Removing the dependency on threshold stability control points provides more stable results and can also result in less time in the field. We hope this framework will allow for more efficient processing of the ever-increasing coastal UAV datasets. 
 
 # Summary
 CoastalLens is set up as 4 scripts (with an optional 5th script) run sequentially from a main entry point script (``UAV_rectification.m``). This allows users to execute parts or all of the full framework depending on their workflow. The first script, ``input_day_flight_data.m``, prompts user input and returns all the user-specified required input data organized in stuctures to be used by the subsequent scripts. The user is required to input data for each day and flight to process. Required user inputs are the video timezone, camera intrinsics, the Products (types of images) to be generated (e.g. Grid/Rectified Image, xTransect or yTransect), and the ground control points to determine the camera world position (via GPS points or pointcloud) ([@hartley:2004], [@gao:2003], [@conlin:2020]). Users can load in pre-set values for the relevant day-specific information from a configuration file. This is useful if similar UAV missions are flown repeatedly at the same location. The second script, ``extract_images_from_UAV.m``, extracts images from the video files at the specified frame rates. This is done via a system command to the ffmpeg command line tool. 
