### 2024-09-17 e586904
Update paper.bib
