+++ paper/paper.bib
@@ -1,4 +1,27 @@
 
+@article{wang:2023,
+	title = {Video stabilization: A comprehensive survey},
+	volume = {516},
+	url = https://www.sciencedirect.com/science/article/abs/pii/S092523122201270X},
+	doi = {10.1016/j.neucom.2022.10.008},
+	pages = {205--230},
+	journaltitle = {Neurocomputing},
+	author = {Wang, Yiming and Huang, Qian and Jiang, Chuanxu and Liu, Jiwen and Shang, Mingzhou and Miao, Zhuang},
+	date = {2023-01},
+	langid = {english},
+}
+
+@article{guilluy:2021,
+	title = {Video stabilization: Overview, challenges and perspectives},
+	volume = {90},
+	url = {https://www.sciencedirect.com/science/article/abs/pii/S0923596520301697},
+	doi = {10.1016/j.image.2020.116015},
+	journaltitle = {Signal Processing: Image Communication},
+	author = {Guilluy, Wilko and Oudre, Laurent and Beghdadi, Azeddine},
+	date = {2021 - 01},
+	langid = {english},
+}
+
 @article{brown:2007,
 	title = {Automatic Panoramic Image Stitching using Invariant Features},
 	volume = {74},
