### 2024-09-17 efde444
Update paper.bib
