+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @article{wang:2023,
 	title = {Video stabilization: A comprehensive survey},
 	volume = {516},
-	url = https://www.sciencedirect.com/science/article/abs/pii/S092523122201270X},
+	url = {https://www.sciencedirect.com/science/article/abs/pii/S092523122201270X},
 	doi = {10.1016/j.neucom.2022.10.008},
 	pages = {205--230},
 	journaltitle = {Neurocomputing},
