+++ paper/paper.bib
@@ -6,7 +6,7 @@
 	doi = {10.1016/j.neucom.2022.10.008},
 	pages = {205--230},
 	journaltitle = {Neurocomputing},
-	author = {Wang, Yiming and Huang, Qian and Jiang, Chuanxu and Liu, Jiwen and Shang, Mingzhou and Miao, Zhuang},
+	author = {{Yiming Wang} and {Qian Huang} and {Chuanxu Jiang} and {Jiwen Liu} and {Mingzhou Shang} and {Zhuang Miao}},
 	date = {2023-01},
 	langid = {english},
 }
@@ -17,8 +17,8 @@
 	url = {https://www.sciencedirect.com/science/article/abs/pii/S0923596520301697},
 	doi = {10.1016/j.image.2020.116015},
 	journaltitle = {Signal Processing: Image Communication},
-	author = {Guilluy, Wilko and Oudre, Laurent and Beghdadi, Azeddine},
-	date = {2021 - 01},
+	author = {{Wilko Guilluy} and {Laurent Oudre} and {Azeddine Beghdadi}},
+	date = {2021-01},
 	langid = {english},
 }
 
