### 2024-09-17 #7111 @AthinaLange
@chickadel I see. I added in a sentence in the intro that mentions that video stabilization is a wide field and included references to two recent review articles.
