### 2024-09-20 1ba1926
Update define_xtransect.m
