### 2024-09-30 04f430e
Update run_cBathy.m
added 'flights' variable catch
