### 2024-09-30 8fc4cb6
Update run_cBathy.m
fix flights var
