+++ CODES/cBathy/run_cBathy.m
@@ -144,10 +144,10 @@ for  dd = 1 : length(day_files)
 
     switch camera_type
         case 'UAV'
-	    try
-	        load(fullfile(day_files(dd).folder, day_files(dd).name, 'day_config_file.mat'), 'flights')
-	    catch
-	        flights = dir(fullfile(day_files(dd).folder, day_files(dd).name)); flights([flights.isdir]==0)=[];
+	    
+	    load(fullfile(day_files(dd).folder, day_files(dd).name, 'day_config_file.mat'), 'flights')
+	    if ~exist('flights', 'var')
+		flights = dir(fullfile(day_files(dd).folder, day_files(dd).name)); flights([flights.isdir]==0)=[];
 	        flights(contains({flights.name}, '.'))=[]; flights(contains({flights.name}, 'GCP'))=[];
 	    end
             assert(exist('flights', 'var'), 'Error (run_cBathy): flights must exist and be stored in ''day_config_file.mat''.')
