+++ DATA/readme.txt
@@ -1,0 +1,3 @@
+DEM.mat should be placed in 20211215_Torrey folder.
+Create folder 'Timestacks' in 20211215_Torrey/Flight_04 folder. This will be your 'Timestacks repository'.
+Create folder 'Timestacks' and 'Predictions' in 20211215_Torrey/Flight_04/Timestacks folder.
