+++ CODES/cBathy/cBathy_v2.0/csmInvertKAlpha.m
@@ -1,254 +1,0 @@
-function [fDependent, camUsed] = ...
-    csmInvertKAlpha(f, G, xy, cam, xm, ym, bathy)
-%
-% [fDependent, camUsed] = ...
-%   csmInvertKAlpha(f, G, xy, cam, xm, ym, bathy)
-%
-% estimate wavennumber components from cross spectral matrix CSM
-% using local nolinear model for any provided subset of Fourier values
-% at any location, xm, ym.    
-%
-% Input
-%   f   frequencies of input Fourier coefficients, Nf x 1
-%   G, Fourier coeffients from a tiles worth of pixels, Nxy x 1 (complex)
-%   xy, pixel locations (Nxy of them); horizontal position only
-%   cam, camera number for each xyz, required for handling seams
-%   xm, ym, single location for this analysis
-%   bathy, building results structure with everything in it.
-% Outputs:
-%   fDependent: small structure of results with the following fields, each
-%   of length nKeep (one for each frequency)
-%       - fB,       - analysed frequencies
-%       - k a h,    - wavenumber, wave direction, depth
-%       - kEst aEst hEst - errors in wavenumber, direction and depth
-%       - skill dof, - skill and degrees of freedom
-
-% 03/07/10 - Holman fix to catch edge error where a tomographic point
-%   an extrapolation region can get poor estimates with low error
-%   because only a very few pixels are fit well.  Fix is to require
-%   nGood pixels (likely > 3) and to have max(deltax) and max(deltay)
-%   at least wavelengthFraction of the expected wavelength.
-% 09/01/11 - Holman rewrite to finally get rid of tiling and analyze a
-%   single location only
-% 12/05/11 - Holman major change to base on EOFs rather than direct from CSM
-
-% 05/22/17 - Holman major change to dynamic tile size (based on fB) and new
-%               seed based on initial guess at bathy, bathy.h0
-% 07/24/19 - major changes including better implementation of dynamic tile
-%            size, removing phi0 as a search variable and new seed
-%            algorithm based on radon transform.
-
-%% %%  1. constants and parameters
-
-global callCount            % keep track of number of calls to predictCSM
-global centerInd            % index of pixel closest to xm, ym for phase corr.
-global v                    % dominant eigenvector at each f
-% set up nonlinear solver parameters.
-
-OPTIONS = statset('nlinfit');   % fit options
-OPTIONS.MaxIter = 50;
-OPTIONS.TolFun = 1e-5;      % previous version used 1e-6
-%OPTIONS.Display = 'iter';  % only for debugging search
-warning off
-
-clear fDependent
-g=9.81;                         % 'g'
-ri = [0:0.01:1]';       % create a hanning filter for later interp
-ai = (1 - cos( pi*(0.5 + 0.5*ri) ).^2);
-if( cBDebug( bathy.params, 'DOPLOTPHASETILE' ))
-    for i = 1: bathy.params.nKeep
-        figure(i); clf
-    end
-end
-
-% find the dominant frequencies for this location, estimates of
-% seed for k and alpha, create adaptive tiles subG, subxy and
-% subCams.  If insufficient pixels, return empty results, indicated by fs 
-% being nan.
-[fs, kAlpha0, subvs, subxy, camUsed, lam1Norms, centerInds] = ...
-                    prepareTiles(f, G, xy, cam, xm, ym, bathy);
-fDependent.fB = fs;
-
-%% %%  Loop through frequencies from most to least coherent solving for the
-%      best fit kAlpha from the complex eigenvectors for each tile
-
[…]
