+++ CODES/cBathy/cBathy_v2.0/cBathyVersion.m
@@ -1,36 +1,0 @@
-function version = cBathyVersion()
-
-%  version = cBathyVersion();
-%
-%  return the version number of cBathy, to be saved with the cBathy
-%  struct.
-
-% adds seam fix from OSU. Also new initialization function to improve
-% initial guesses.
-version = 2.0;
-
-
-%
-%   Copyright (C) 2017  Coastal Imaging Research Network
-%                       and Oregon State University
-
-%    This program is free software: you can redistribute it and/or  
-%    modify it under the terms of the GNU General Public License as 
-%    published by the Free Software Foundation, version 3 of the 
-%    License.
-
-%    This program is distributed in the hope that it will be useful,
-%    but WITHOUT ANY WARRANTY; without even the implied warranty of
-%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-%    GNU General Public License for more details.
-
-%    You should have received a copy of the GNU General Public License
-%    along with this program.  If not, see
-%                                <http://www.gnu.org/licenses/>.
-
-% CIRN: https://coastal-imaging-research-network.github.io/
-% CIL:  http://cil-www.coas.oregonstate.edu
-%
-%key cBathy
-%
-
