+++ CODES/cBathy/run_cBathyCT.m
@@ -1,405 +1,0 @@
-%% run_cBathyCT
-% run_cBathyCT returns cBathyCT output based off Lange et al. 2023
-%
-%% Description
-%
-%   Args:
-%           global_dir (string) : global directory - where CODES and (typically) DATA  are located.
-%           day_files (structure) : folders of the days to process - requires day_files.folder and day_files.name
-%           camera_type (string) : 'ARGUS' or 'UAV' depending on processing
-%           flights (structure) : for camea_type = 'UAV' : folders of the flights to process - requires flights.folder and flights.name
-%           cam_num (double) : for camea_type = 'ARGUS' : number of cameras to loop through
-%           Products (structure) : Data products
-%                       type (string) : 'Grid', 'xTransect', 'yTransect'
-%                       lat (double) : latitude of origin grid
-%                       lon (double): longitude of origin grid
-%                       xlim (double): [1 x 2] cross-shore limits of grid (+ is offshore of origin) (m)
-%                       ylim (double) : [1 x 2] along-shore limits of grid (+ is to the right of origin looking offshore) (m)
-%                       dx (double) : Cross-shore resolution (m)
-%                       dy (double) : Along-shore resolution (m)
-%                       x (double): Cross-shore distance from origin (+ is offshore of origin) (m)
-%                       y (double): Along-shore distance from origin (+ is to the right of the origin looking offshore) (m)
-%                       z (double) : Elevation - can be empty or array of DEM values (NAVD88 m)
-%                       tide (double) : Tide level (NAVD88 m)%                       Eastings (double) : [y x x] Eastings coordinates (m)
-%                       Northings (double) : [y x x] Northings coordinates (m)
-%                       Irgb_2d (uint8 image) : [m x y_length x x_length x 3] timeseries of pixels extracted according to dimensions of xlim and ylim
-%           bathy (structure) : cBathy output after Step 2
-%                       fCombined (structure) : cBathy Step 2
-%                                   h (array) : [y_length x x_length] cBathy depth estimate
-%                                   hErr (array) : [y_length x x_length] cBathy error estimates
-%                       coords (structure) :
-%                                   Eout (array) :  [y_length x x_length] cross-shore location in Eastings
-%                                   Nout (array) :  [y_length x x_length] cross-shore location in Northings
-%           DEM (structure) : Digital Elevation Map Data - define_DEM.m
-%                       time (datetime) : date of jumbo survey used
-%                       type (cell array) : Survey type for each transect
-%                       MOP (cell array) : Mop number of each x transect
-%                       X (array): [mop_length x cross-shore length] - Eastings
-%                       Y (array): [mop_length x cross-shore length] - Northings
-%                       Z (array): [mop_length x cross-shore length] - Elevation (NAVD88m)
-%
-%
-%   Returns:
-%           cBathyCT (structure) : cBathyCT results
-%                       date (string) : date of product (YYYYMMDD)
-%                       Iocation (string) : location name - from filename
-%                       flight / cam (double) : flight number (for UAV) or camera number (for ARGUS)
-%                       mop (double) : mop number
-%                       x10 (array) : [0:0.1:500], 0.1m resolution - all elevation values interpolated to this x array
[…]
