### 2024-10-01 #7111 @chickadel
@Kevin-Mattheus-Moerman I'm okay with all the written parts of the paper.  However, I can;t verify the demo works, and it's frankly more than I have time for to diagnose where it's breaking exactly (currently it can't upload the image data, which seems benign).  I'm fine if the other reviewer (@kabernagelm) has verified that this code and demo work as intended.
