### 2024-10-02 #7111 @AthinaLange
@chickadel Did you download the video from [zenodo](https://zenodo.org/records/10810778)? After running 'extract_images_from_UAV.m' you should have a folder called 'images_1Hz' inside of 'DATA/20211215_Torrey/Flight_04'.
The 62 images that should have been extracted should then be listed in the imageDatastore variable (line 105 of 'stabilize_video.m').

running this command, you should see that images.Files has the full path name of all the images:
`images = imageDatastore(fullfile(data_dir, '20211215_Torrey', 'Flight_04', 'images_1Hz'));`
