### 2024-10-15 667a2e6
Update UAV_rectification_DEMO.m
