### 2024-10-15 d0a9563
Update UAV_rectification_DEMO.m
