### 2024-10-20 #7111 @Kevin-Mattheus-Moerman
@AthinaLange the above list contains the last set of required steps. Please create a new release and archive the new release on ZENODO. You can create the archive manually or use the ZENODO-GitHub integration (e.g. see instructions like this https://coderefinery.github.io/github-without-command-line/doi/). Please let me know what the version tag is for your new release.

I checked the paper and did not see spelling/type issues, however I do have this comment:
- [ ] Please spell out USA as `United States of America` in your affiliations.
