### 2024-10-20 58776d8
Update get_products_DEMO.m
