### 2024-10-20 62961f1
Update get_products.m
