### 2024-10-21 #7111 @AthinaLange
I've changed the affiliations and made a new release v1.1 and it looks like it is linked in the zenodo archive (https://zenodo.org/records/13958678)
