### 2024-10-24 #7111 @Kevin-Mattheus-Moerman
@AthinaLange thanks for those updates. Now can you please manually edit the ZENODO archive listing so that:

- [ ] The author set is the same as the paper
- [ ] The title matches the paper title
- [ ] The license matches your software license (see Rights which currently shows a default cc-by license).
