### 2024-11-21 #7111 @AthinaLange
I'm having issues accessing the Zenodo account with the release, but it should be solved next week, then I will update it right away.
