### 2024-12-03 #7111 @Kevin-Mattheus-Moerman
@AthinaLange thanks for working on those steps. The following are still open:

- [ ] The archive title should match the paper title in full (`CoastalLens: A MATLAB UAV Video Stabilization & Rectification Framework`)
- [ ] The author order should ideally match the paper author order. Further more the formatting of the author names differs. The first author should include the comma and have the format `Lastname, Firstname`, and so become `Lange, Athina`. In addition the author listed as `jwfiedler` should instead be listed as `Fiedler, Julia W`.
- [ ] At this point you may also consider adding ORCIDs to the ZENODO archive (not a requirement but recommended).
