### 2024-12-04 #7111 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
