### 2024-12-04 #7111 @Kevin-Mattheus-Moerman
@AthinaLange as AEiC for JOSS I will now help to process this submission for acceptance in JOSS.
